Made changes to the css that caused the page to break.
