+++ src/components/common-components/top-nav.css
@@ -1,13 +1,12 @@
 .main-container {
     display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 35px;
-    color: #003371;
+    flex-direction: row;
+    max-height: 35px;
+    background-color: #333;
 }
+// the max-height made it to look like this
 
-.top-nav {
+/* .top-nav {
     width: 100%;
     height: 35px !important;
     background-color: #333;
@@ -17,10 +16,17 @@
     justify-content: space-between;
     padding: 0 10px;
     box-sizing: border-box;
+} */
+
+.logo-container {
+    width: 40% !important;
+}
+
+.title-container {
+    width: 60% !important;
 }
 
 .top-nav a {
-    color: white;
     text-decoration: none;
     padding: 0 10px;
 }
@@ -30,6 +36,7 @@
 }
 
 ul {
+    width: 80%;
     list-style: none;
     padding: 0;
 }
