+++ src/components/questions-page/question-card.css
