+++ src/screens/questions.js
@@ -1,20 +1,15 @@
 import { Link } from "react-router-dom";
 import React from "react";
-import { TopNav } from "../components/top-nav";
-import styles from "../styles/questions.css";
-import QuestionCard from "../components/question-card";
+import { TopNav } from "../components/common-components/top-nav";
+import "../styles/page-styles/questions.css";
+import QuestionCard from "../components/questions-page/question-card";
 
 export function Questions() {
     return (
         <div>
-            <div className="styles.main-container">
-                <TopNav />
-                <h1>This is the questions page!</h1>
-                <p>
-                    <Link to="/dashboard">Go to Dashboard</Link>
-                </p>
-
-                <div className="styles.question-container">
+            <TopNav />
+            <div className="main-container">
+                <div className="question-container">
                     <QuestionCard />
                     <QuestionCard />
                     <QuestionCard />
