--- tailwind.config.js
@@ -1,8 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-    content: ["./src/**/*.{html,js,jsx,ts,tsx}"],
-    theme: {
-        extend: {},
-    },
-    plugins: [],
-};
